:star2: pass dredd and all cucumber and mocha tests :rainbow: close #13
