+++ config/test.json
@@ -5,10 +5,11 @@
   "mongoose": {
     "debug": false
   },
-  "mailer": {
+  "transport": {
     "mailgun": {
-      "batch_limit": 1000,
-      "custom_batch_limit": 2
+      "apiKey": "",
+      "domain": "",
+      "batch_limit": 2
     }
   }
 }
