+++ cucumber/features/channel_features/channel_create.feature
@@ -22,4 +22,3 @@ Feature: A requester asks for the creation of a Channel object
     | 01f0000000000000003f0002 | /api/channel  | channel/valid_channel_name_two   | 201    |
     | 01f0000000000000003f0003 | /api/channel  | channel/duplicated_channel_name  | 409    |
     | 01f0000000000000003f0001 | /api/channel  | channel/empty_channel_object     | 400    |
-    | 01f0000000000000003fasdf | /api/channel  | channel/valid_channel_name_one   | 400    |
