+++ cucumber/features/step_definitions/email_steps.js
@@ -1,4 +1,3 @@
-var assert = require('assert');
 var config = require('config');
 var nock = require('nock');
 
@@ -29,14 +28,6 @@ module.exports = function() {
         request
             .send(emailObj)
             .expect(res.status)
-            .end(function(err, response) {
-
-            if (err) {
-                return callback(err);
-            }
-            assert.deepEqual(response.body, res.data, 'Responses do not match');
-
-            return callback();
-        });
+            .end(callback);
     });
 };
