+++ cucumber/test_files/email/valid_email_response.json
@@ -1,6 +1,3 @@
 {
-  "status": 200,
-  "data": {
-    "output": "Queued notifications"
-  }
+  "status": 204
 }
