+++ cucumber/test_files/sms/valid_sms_response.json
@@ -1,6 +1,3 @@
 {
-  "status": 200,
-  "data": {
-    "output": "Queued notifications"
-  }
+  "status": 204
 }
