+++ doc/api/channel.md
@@ -12,7 +12,7 @@ Retrieves the list of Channel objects associated to the requester's Identity obj
 
     + Headers
 
-            Authorization: Bearer BEWkwDA0bTTw_4dSPNI8lDPWF
+            x-user-id: 01f0000000000000003f0001
 
 
 + Response 200 (application/json; charset=utf-8)
@@ -43,7 +43,7 @@ Retrieves the list of Channel objects associated to the requester's Identity obj
 
 Creates a new Channel object with the provided name in the data store
 
-+ Request
++ Request (application/json)
 
     + Body
 
@@ -75,18 +75,14 @@ Creates a new Channel object with the provided name in the data store
 Replaces the data of the Channel object identified by the provided channel Id.
 
 + Parameters
-    + id (required, string, `12345`)
+    + id (required, string, `01f0000000000000006f0001`)
 
         The ID of the Channel object to update
 
 + Attributes
-    + id: 12345 (string)
+    + id: 01f0000000000000006f0001 (string)
 
-+ Request
-
-    + Headers
-
-            Authorization: Bearer BEWkwDA0bTTw_4dSPNI8lDPWF
++ Request (application/json)
 
     + Body
 
@@ -110,7 +106,7 @@ Replaces the data of the Channel object identified by the provided channel Id.
 Delete the Channel object identified by the provided channel Id.
 
 + Parameters
-    + id (required, string, `12345`) ... The ID of the Channel object to delete
+    + id (required, string, `01f0000000000000006f0001`) ... The ID of the Channel object to delete
 
 + Request
 
@@ -136,13 +132,9 @@ Delete the Channel object identified by the provided channel Id.
 Retrieves the list of Identity objects associated to the specified Channel object
 
 + Parameters
-    + id (required, string, `12345`) ... The ID of the Channel object to retrieve data from
+    + id (required, string, `01f0000000000000006f0001`) ... The ID of the Channel object to retrieve data from
 
 + Request
-
-    + Headers
-
-            Authorization: Bearer BEWkwDA0bTTw_4dSPNI8lDPWF
 
 + Response 200
 
@@ -189,14 +181,10 @@ Retrieves the list of Identity objects associated to the specified Channel objec
 Delete an Identity identifier from a Channel object
 
 + Parameters
-    + channelId (required, string, `12345`) ... The Channel identifier from which the Identity shall be removed
-    + identityId (required, string, `5678`) ... The Identity identifier that shall be removed from the Channel object
+    + channelId (required, string, `01f0000000000000006f0001`) ... The Channel identifier from which the Identity shall be removed
+    + identityId (required, string, `01f0000000000000003f0001`) ... The Identity identifier that shall be removed from the Channel object
 
 + Request
-
-     + Headers
-
-            Authorization: Bearer BEWkwDA0bTTw_4dSPNI8lDPWF
 
 + Response 204
