+++ doc/api/email.md
@@ -20,17 +20,16 @@ __message__ | String | Required | Email Text Message to be sent.
     + Body
 
             {
-              "to": ["01f0000000000000003f0002"],
-              "from": "noreply@email.com",
-              "message": "This is the body of the email"
+              "identities": ["01f0000000000000003f0001", "01f0000000000000003f0002", "01f0000000000000003f0003"],
+              "channels": ["buddies"],
+              "content": {
+                "from": "noreply@email.com",
+                "message": "This is the body of the email"
+              }
             }
 
-+ Response 201 (application/json; charset=utf-8)
++ Response 204
 
-            {
-              "id": "<20150729090542.29416.27304@sandboxd41385ca8b5f40eda0c863e41dc6e997.mailgun.org>",
-              "message": "Queued. Thank you."
-            }
 
 + Response 400 (application/json; charset=utf-8)
