+++ doc/api/sms.md
@@ -20,16 +20,14 @@ __message__ | String | Required | Sms Text Message to be sent.
     + Body
 
             {
-                "to": [
-                    "01f0000000000000003f0002",
-                    "01f0000000000000003f0003"
-                ],
-                "from": "+15005550006",
-                "message": "Hello there!"
+                "identities": [],
+                "channels": ["buddies"],
+                "content": {
+                    "from": "+15005550010",
+                    "message": "Hola holita RESONATOR"
+                }
             }
 
-+ Response 201 (application/json)
-
-             {}
++ Response 204
