+++ lib/middleware/check_channel.js
@@ -3,7 +3,7 @@ var _ = require('lodash');
 var errors = require('../util/errors');
 
 module.exports = function() {
-  return function check_channel(req, res, next) {
+  return function checkChannel(req, res, next) {
 
     var channelObj = req.body;
