+++ lib/middleware/check_url.js
@@ -8,7 +8,7 @@ module.exports = function() {
     var method = req.method;
 
     req.isPublic = _.some(config.get('auth.excludeAuthPaths'), function(pattern) {
-      return path.match(new RegExp(pattern.endpoint, 'g')) && (pattern.method === method);
+      return path.match(new RegExp(pattern.endpoint, 'g')) && (!pattern.method || pattern.method === method);
     });
 
     return next();
