+++ lib/platforms/channel.js
@@ -29,7 +29,7 @@ function retrieveChannelDataForIdentity(identity, callback) {
     Channel.find().where('name').in(foundIdentity.channels).exec(function(err, foundChannels) {
 
       if (err) {
-        return callback(new errors.InternalError('Could not fetch the channels associated to the provided Identity object'));
+        return callback(new errors.BadRequestError('Could not fetch the channels associated to the provided Identity object'));
       }
 
       var channelData = [];
@@ -214,7 +214,7 @@ function deleteIdentityFromChannel(channelId, identityId, callback) {
       getChannel(channelId, function(err, foundChannel) {
 
         if (err) {
-          return done(new errors.InternalError('Could not fetch requested Channel object'));
+          return done(new errors.BadRequestError('Could not fetch requested Channel object'));
         }
 
         if (_.isEmpty(foundChannel)) {
@@ -229,7 +229,7 @@ function deleteIdentityFromChannel(channelId, identityId, callback) {
       identities.get(identityId, function(err, foundIdentity) {
 
         if (err) {
-          return done(new errors.InternalError('Could not fetch requested Identity object'));
+          return done(new errors.BadRequestError('Could not fetch requested Identity object'));
         }
 
         if (_.isEmpty(foundIdentity)) {
@@ -264,23 +264,9 @@ function deleteIdentityFromChannel(channelId, identityId, callback) {
         deleteChannelIdFromIdentity: function(microDone) {
           identities.removeValuesFromField(identityId, 'channels', channel.name, microDone);
         }
-      }, function(err) {
+      }, done);
 
-        if (err) {
-          return done(new errors.InternalError('Deletion process failed'));
-        }
-
-        return done();
-      });
-    }], function finishDeletion(err) {
-
-      if (err) {
-        return callback(new errors.InternalError('Could not delete provided Identity object from channel'));
-      }
-
-      return callback();
-    }
-  );
+    }], callback);
 }
 
 function removeValuesFromField(channelList, field, valuesToRemove, callback) {
