+++ lib/platforms/email.js
@@ -1,9 +1,8 @@
-var request = require('request');
 var config = require('config');
 var _ = require('lodash');
 var log = require('../util/logger');
+var transport = require('../transport/mailgun');
 
-var MAILGUN_MESSAGES_URL = 'https://api.mailgun.net/v3/' + config.get('transport.mailgun.domain') + '/messages';
 var BATCH_LIMIT = config.get('platform_batch_limits.mailgun');
 var CUSTOM_BATCH_LIMIT = config.get('transport.mailgun.batch_limit') || BATCH_LIMIT;
 
@@ -13,45 +12,25 @@ function sendEmailNotification(identities, body, callback) {
     return identity.devices.email;
   }));
 
-      // Format email object
-      var data = {
-        to: emails,
-        from: body.content.from,
-        subject: body.content.subject,
-        html: body.content.message
-      };
+  log.info('Sending request to send emails');
 
-      log.info('Sending request to send emails');
-      log.debug('Sending request to send emails', data);
+  transport.send(emails, body.content, function (err, response, body) {
 
-      request({
-        url: MAILGUN_MESSAGES_URL,
-        method: 'POST',
-        json: true,
-        auth: {
-          user: 'api',
-          pass: config.get('transport.mailgun.apiKey'),
-          sendImmediately: true
-        },
-        formData: data
+    if (err) {
+      log.error('Error sending request to send emails', err);
+      return callback(err);
+    }
 
-      }, function (err, response, body) {
+    var output = {
+      response: response,
+      body: body
+    };
 
-        if (err) {
-          log.error('Error sending request to send emails', err);
-          return callback(err);
-        }
+    log.info('Receiving response from email request');
+    log.debug('Receiving response from email request', body);
 
-        var output = {
-          response: response,
-          body: body
-        };
-
-        log.info('Receiving response from email request');
-        log.debug('Receiving response from email request', body);
-
-        return callback(null, output);
-      });
+    return callback(null, output);
+  });
 }
 
 var emailOptions = {
