+++ lib/platforms/push.js
@@ -1,8 +1,8 @@
 var async = require('async');
 var config = require('config');
 var _ = require('lodash');
-var gcm = require('../util/gcm');
-var apn = require('../util/apn');
+var gcm = require('../transport/gcm');
+var apn = require('../transport/apn');
 var Identity = require('../models/identity');
 var log = require('../util/logger');
