+++ lib/platforms/sms.js
@@ -1,21 +1,12 @@
-var twilio = require('twilio');
 var async = require('async');
 var config = require('config');
 var _ = require('lodash');
 var log = require('../util/logger');
 
-var TWILIO_ACCOUNT_SID = config.get('transport.twilio.account_sid');
-var TWILIO_AUTH_TOKEN = config.get('transport.twilio.auth_token');
+var transport = require('../transport/twilio');
 
 var BATCH_LIMIT = config.get('platform_batch_limits.twilio');
 var CUSTOM_BATCH_LIMIT = config.get('transport.twilio.batch_limit') || BATCH_LIMIT;
-
-if (process.env.NODE_ENV === 'test') {
-    TWILIO_ACCOUNT_SID = "Some_random_account_sid";
-    TWILIO_AUTH_TOKEN = "Some_random_auth_token";
-}
-
-messager = new twilio.RestClient(TWILIO_ACCOUNT_SID, TWILIO_AUTH_TOKEN);
 
 function sendSmsNotification(identities, body, callback) {
 
@@ -36,7 +27,7 @@ function sendSmsNotification(identities, body, callback) {
     log.info('Sending request to send sms');
     log.debug('Sending request to send sms', sms);
 
-    messager.sendSms(sms, function (err) {
+    transport.sendSms(sms, function (err) {
 
       if (err) {
         return done(err);
