+++ lib/routes/channel.js
@@ -21,7 +21,7 @@ routes.getChannel = function(req, res, next) {
   var identity = req.identity;
 
   if (!identity) {
-    return next(new errors.BadRequestError('Unknown provided channel identifier'));
+    return next(new errors.BadRequestError('Unknown provided identity'));
   }
 
   channels.retrieveChannelDataForIdentity(identity, function(err, channelData) {
@@ -36,12 +36,6 @@ routes.getChannel = function(req, res, next) {
 };
 
 routes.createChannel = function(req, res, next) {
-
-  var identity = req.identity;
-
-  if (!identity) {
-    return next(new errors.BadRequestError('Unknown provided channel identifier'));
-  }
 
   channels.createChannel(req.body, function(err, createdChannel) {
