+++ lib/routes/email.js
@@ -1,5 +1,4 @@
 var emailPlatform = require('../platforms/email');
-var errors = require('../util/errors');
 var checkEmail = require('../middleware/check_email');
 var orchestrator = require('../platforms/orchestrator');
 
@@ -12,12 +11,6 @@ module.exports = function(server) {
 
 routes.sendEmailNotification = function(req, res, next) {
 
-    var identity = req.identity;
-
-    if (!identity) {
-        return next(new errors.BadRequestError('Unknown identity'));
-    }
-
     orchestrator.sendNotification(req.body, emailPlatform, function(err, output) {
 
         if (err) {
@@ -25,7 +18,7 @@ routes.sendEmailNotification = function(req, res, next) {
         }
 
         req.log.info({res: output.response}, 'mail_sent');
-        res.send(200, output.body);
+        res.send(204);
         return next();
     });
 };
