+++ lib/routes/identity.js
@@ -8,7 +8,6 @@ module.exports = function(server) {
   server.put('/api/identity', routes.updateIdentity);
   server.post('/api/identity', routes.createIdentity);
   server.get('/api/identity/:identityId', routes.getIdentityById);
-  server.put('/api/identity/:identityId', routes.updateIdentityById); // This one does not make sense, or does it?
 
 };
 
@@ -84,9 +83,3 @@ routes.getIdentityById = function(req, res, next) {
   });
 };
 
-routes.updateIdentityById = function(req, res, next) {
-
-  // Get identity from params
-  // Update properties of identity via manager
-  return next();
-};
