+++ lib/routes/push.js
@@ -1,5 +1,4 @@
 var orchestrator = require('../platforms/orchestrator');
-var errors = require('../util/errors');
 var checkPushNotification = require('../middleware/check_push_notification');
 var pushPlatform = require('../platforms/push');
 var routes = {};
@@ -12,19 +11,13 @@ module.exports = function(server) {
 
 routes.sendPushNotification = function(req, res, next) {
 
-    var identity = req.identity;
-
-    if (!identity) {
-        return next(new errors.BadRequestError('Unknown identity'));
-    }
-
-    orchestrator.sendPushNotifications(req.body, pushPlatform, function(err, response) {
+    orchestrator.sendPushNotifications(req.body, pushPlatform, function(err /*, response */) {
 
         if (err) {
             return next(new errors.InternalError('Could not send PUSH Notification to destination'));
         }
 
-        res.send(200, response.body);
+        res.send(204);
         return next();
     });
 };
