+++ lib/routes/sms.js
@@ -1,5 +1,4 @@
 var smsPlatform = require('../platforms/sms');
-var errors = require('../util/errors');
 var checkSms = require('../middleware/check_sms');
 var orchestrator = require('../platforms/orchestrator');
 
@@ -12,19 +11,13 @@ module.exports = function(server) {
 
 routes.sendSmsNotification = function(req, res, next) {
 
-    var identity = req.identity;
-
-    if (!identity) {
-        return next(new errors.BadRequestError('Unknown identity'));
-    }
-
-    orchestrator.sendNotification(req.body, smsPlatform, function(err, response) {
+    orchestrator.sendNotification(req.body, smsPlatform, function(err/*, response */) {
 
         if (err) {
             return next(new errors.InternalError('Could not send SMS to destination'));
         }
 
-        res.send(200, response.body);
+        res.send(204);
         return next();
     });
 };
