+++ lib/service.js
@@ -14,7 +14,8 @@ var ROUTES_FOLDER = process.cwd() + '/lib/routes/';
 
     server.use(restify.bodyParser());
     server.use(function (req, res, next) {
-      console.log('> ' + req.method + ' ' + req.url);
+      res.charSet('utf-8');
+      log.info('> ' + req.method + ' ' + req.url);
       next();
     });
 
@@ -22,14 +23,13 @@ var ROUTES_FOLDER = process.cwd() + '/lib/routes/';
       var timing = Date.now() - new Date(req._time);
 
       if(res._data){
-        console.log('< ', res.statusCode, res._data.length, 'bytes', timing, 'ms');
+        log.info('< ', res.statusCode, res._data.length, 'bytes', timing, 'ms');
       } else {
-        console.log('< ', res.statusCode, 'empty response', timing, 'ms');
+        log.info('< ', res.statusCode, 'empty response', timing, 'ms');
       }
     });
 
     server.on('uncaughtException', function (req, res, router, error) {
-      console.log(error.stack());
       log.error('UncaughtException', error);
     });
